Move ExcThread class to meta_class package
Also add type hinting and fix mutable arguments
